+++ src/imio/smartweb/core/webcomponents/src/components/Events/ContactContent/ContactContent.jsx
@@ -4,6 +4,8 @@ import useAxios from "../../../hooks/useAxios";
 import useFilterQuery from "../../../hooks/useFilterQuery";
 import moment from "moment";
 import Moment from "react-moment";
+import ReactMarkdown from 'react-markdown'
+
 const ContactContent = ({ queryUrl, onChange }) => {
     let history = useHistory();
     const queryString = require("query-string");
@@ -327,7 +329,9 @@ const ContactContent = ({ queryUrl, onChange }) => {
 						)}
 					</div>
 				</div>
-				<p className="r-content-description">{contactItem.description}</p>
+				<div className="r-content-description">
+					<ReactMarkdown>{contactItem.description}</ReactMarkdown>
+				</div>
 				<div
 					className="r-content-text"
 					dangerouslySetInnerHTML={{