+++ CHANGES.rst
@@ -2,7 +2,7 @@ Changelog
 =========
 
 
-1.0.10 (unreleased)
+1.0.10 (2022-05-10)
 -------------------
 
 - Add description for directory items